[marlowe-integration-tests] flipped file structure to reflect domain
