+++ marlowe-integration-tests/test/Language/Marlowe/Runtime/Web/Contracts/Contract/GetContract.hs
@@ -1,4 +1,4 @@
-module Language.Marlowe.Runtime.Web.GetContract where
+module Language.Marlowe.Runtime.Web.Contracts.Contract.GetContract where
 
 import Control.Exception (throw)
 import Control.Monad.IO.Class (MonadIO(liftIO))
