+++ marlowe-integration-tests/test/Language/Marlowe/Runtime/Web/Contracts/GetContracts.hs
@@ -1,4 +1,4 @@
-module Language.Marlowe.Runtime.Web.GetContracts where
+module Language.Marlowe.Runtime.Web.Contracts.GetContracts where
 
 import Control.Exception (throw)
 import Control.Monad.IO.Class (MonadIO(liftIO))
