+++ marlowe-integration-tests/test/Language/Marlowe/Runtime/Web/Contracts/Transactions/Transaction/GetTransaction.hs
@@ -1,5 +1,4 @@
-{-# OPTIONS_GHC -Wno-unused-imports #-}
-module Language.Marlowe.Runtime.Web.GetTransaction where
+module Language.Marlowe.Runtime.Web.Contracts.Transactions.Transaction.GetTransaction where
 
 import Control.Monad.IO.Class (MonadIO(liftIO))
 import qualified Control.Monad.Reader as Reader
