+++ marlowe-integration-tests/test/Language/Marlowe/Runtime/Web/Contracts/Transactions/Transaction/PutTransaction.hs
@@ -1,4 +1,4 @@
-module Language.Marlowe.Runtime.Web.PostTransaction where
+module Language.Marlowe.Runtime.Web.Contracts.Transactions.Transaction.PutTransaction where
 
 import Control.Monad.IO.Class (MonadIO(liftIO))
 import qualified Data.Map.Strict as Map
@@ -7,21 +7,23 @@ import Data.Time (getCurrentTime, secondsToNominalDiffTime)
 import Language.Marlowe.Core.V1.Semantics.Types (Input(NormalInput), InputContent(IDeposit))
 import Language.Marlowe.Extended.V1 (ada)
 import Language.Marlowe.Runtime.Integration.Common
+  (Wallet(Wallet, addresses, signingKeys), expectJust, getGenesisWallet, runIntegrationTest, runWebClient)
 import Language.Marlowe.Runtime.Integration.StandardContract (standardContract)
 import Language.Marlowe.Runtime.Plutus.V2.Api (toPlutusAddress)
 import Language.Marlowe.Runtime.Transaction.Api (WalletAddresses(..))
 import Language.Marlowe.Runtime.Web (RoleTokenConfig(RoleTokenSimple))
 import qualified Language.Marlowe.Runtime.Web as Web
-import Language.Marlowe.Runtime.Web.Client (postContract, postTransaction)
-import Language.Marlowe.Runtime.Web.Common (submitContract)
+import Language.Marlowe.Runtime.Web.Client (postContract, postTransaction, putTransaction)
+import Language.Marlowe.Runtime.Web.Common (signShelleyTransaction', submitContract)
 import Language.Marlowe.Runtime.Web.Server.DTO (ToDTO(toDTO))
 import Test.Hspec (Spec, describe, it)
 import Test.Integration.Marlowe.Local (withLocalMarloweRuntime)
 
+
 spec :: Spec
-spec = describe "POST /contracts/{contractId}/transactions" do
+spec = describe "PUT /contracts/{contractId}/transactions/{transaction}" do
   it "returns the transaction header" $ withLocalMarloweRuntime $ runIntegrationTest do
-    partyAWallet <- getGenesisWallet 0
+    partyAWallet@Wallet{signingKeys} <- getGenesisWallet 0
     partyBWallet <- getGenesisWallet 1
 
     result <- runWebClient do
@@ -55,7 +57,7 @@ spec = describe "POST /contracts/{contractId}/transactions" do
 
       let inputs = [NormalInput $ IDeposit partyA partyA ada 100_000_000]
 
-      postTransaction
+      Web.ApplyInputsTxEnvelope{transactionId, txEnvelope} <- postTransaction
         partyAWebChangeAddress
         (Just partyAWebExtraAddresses)
         (Just partyAWebCollataralUtxos)
@@ -68,6 +70,8 @@ spec = describe "POST /contracts/{contractId}/transactions" do
           , inputs
           , tags = mempty
           }
+      applyTx <- liftIO $ signShelleyTransaction' txEnvelope signingKeys
+      putTransaction contractId transactionId applyTx
     case result of
       Left _ ->  fail $ "Expected 200 response code - got " <> show result
-      Right Web.ApplyInputsTxEnvelope{} ->  pure ()
+      Right () ->  pure ()
