+++ marlowe-integration-tests/test/Language/Marlowe/Runtime/Web/Withdrawal/PutWithdrawal.hs
@@ -1,4 +1,4 @@
-module Language.Marlowe.Runtime.Web.PutWithdrawal where
+module Language.Marlowe.Runtime.Web.Withdrawal.PutWithdrawal where
 
 import Control.Monad.IO.Class (MonadIO(liftIO))
 import qualified Data.Set as Set
